Disable race detector for podstatus store tests that use consul.
The version of consul we have vendored in and is used in some tests via
consulutil.NewFixture() contains known data races, so this commit moves
some tests that depend on it to a file that disables the race detector
via build flags.
